chore: raise file_length caps to 450/600 and split LocalHomebrewError out of LocalHomebrewTypes
- Root file_length warning at 450, test override stays 600; test
  type_body_length aligned at 450
- LocalHomebrewError (191 lines, self-contained LocalizedError enum)
  moved to Domain/LocalHomebrewError.swift, LocalHomebrewTypes now 309
- Drop a now-superfluous swiftlint:disable in script/profile_ui.swift

Zero SwiftLint warnings repo-wide.
